# src/docs/swagger.yaml
basePath: /api
definitions:
...
  models.AddBotFriendRequest:
    properties:
      alias:
        type: string
      app_key:
        type: string
      bot_config_id:
        type: integer
      priority:
        type: integer
    required:
    - app_key
    - bot_config_id
    type: object
...
  models.ToggleBotFriendStatusRequest:
    properties:
      is_active:
        type: boolean
    required:
    - is_active
    type: object
...
  models.UpdateBotFriendAppKeyRequest:
    properties:
      app_key:
        type: string
    required:
    - app_key
    type: object
  models.UpdateBotFriendPriorityRequest:
    properties:
      priority:
        type: integer
    required:
    - priority
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: 怒喵 API 文档
  version: "1.0"
paths:
...
  /api/app/friends/bots:
    get:
      consumes:
      - application/json
      description: 获取用户的所有Bot好友
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取Bot好友列表
      tags:
      - 用户好友管理
    post:
      consumes:
      - application/json
      description: 添加Bot为好友
      parameters:
      - description: 好友信息
        in: body
        name: friend
        required: true
        schema:
          $ref: '#/definitions/models.AddBotFriendRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 添加Bot好友
      tags:
      - 用户好友管理
  /api/app/friends/bots/{bot_config_id}:
    delete:
      consumes:
      - application/json
      description: 删除Bot好友关系
      parameters:
      - description: Bot配置ID
        in: path
        name: bot_config_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 好友不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除Bot好友
      tags:
      - 用户好友管理
  /api/app/friends/bots/{bot_config_id}/appkey:
    patch:
      consumes:
      - application/json
      description: 更新Bot好友的LLM API密钥
      parameters:
      - description: Bot配置ID
        in: path
        name: bot_config_id
        required: true
        type: integer
      - description: AppKey信息
        in: body
        name: appkey
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBotFriendAppKeyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 好友不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新Bot好友的AppKey
      tags:
      - 用户好友管理
  /api/app/friends/bots/{bot_config_id}/priority:
    patch:
      consumes:
      - application/json
      description: 更新Bot好友的优先级
      parameters:
      - description: Bot配置ID
        in: path
        name: bot_config_id
        required: true
        type: integer
      - description: 优先级信息
        in: body
        name: priority
        required: true
        schema:
          $ref: '#/definitions/models.UpdateBotFriendPriorityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 好友不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新Bot好友优先级
      tags:
      - 用户好友管理
  /api/app/friends/bots/{bot_config_id}/toggle:
    patch:
      consumes:
      - application/json
      description: 启用或禁用Bot好友
      parameters:
      - description: Bot配置ID
        in: path
        name: bot_config_id
        required: true
        type: integer
      - description: 状态信息
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.ToggleBotFriendStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 好友不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 切换Bot好友状态
      tags:
      - 用户好友管理
...
swagger: "2.0"